refactor: split agent run lifecycle
